requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0
PyYAML>=6.0.0
scikit-learn>=1.4.0
scipy>=1.12.0
statsmodels>=0.14.0
matplotlib>=3.8.0
seaborn>=0.13.0
